deleted: InformationView, InformationRowView, InspectionListView
added: ChapterMainView, FavoriteView, SectionView, SubSectionView
