+++ CrimeFilesApp/ChapterMainView.swift
@@ -1,0 +1,18 @@
+//
+//  ChapterMainView.swift
+//  CrimeFilesApp
+//
+//  Created by Александр Полочанин on 8.11.23.
+//
+
+import SwiftUI
+
+struct ChapterMainView: View {
+    var body: some View {
+        Text("ChapterMainView")
+    }
+}
+
+#Preview {
+    ChapterMainView()
+}
