+++ CrimeFilesApp/ContentView.swift
@@ -1,5 +1,5 @@
 //
-//  ContentView.swift
+//  InformationFiew.swift
 //  CrimeFilesApp
 //
 //  Created by Александр Полочанин on 6.11.23.
@@ -9,45 +9,18 @@ import SwiftUI
 
 struct ContentView: View {
     
-    @StateObject private var viewModel = InspectionViewModel()
+    let title: String
+    let mainText: String
     
     var body: some View {
-        ZStack {
-            NavigationStack {
-                ZStack{
-                    
-//                    MainBackGroundView()
-//                        .ignoresSafeArea()
-
-                    List(viewModel.inspections, id: \.title) { inspectionDetailViewModel in
-                        NavigationLink {
-                            InspectionListView(
-                                inspectionList: inspectionDetailViewModel.rows,
-                                title: inspectionDetailViewModel.title,
-                                imageTitle: inspectionDetailViewModel.titleImage
-                            )
-                            
-                        } label: {
-                            CellView(
-                                title: inspectionDetailViewModel.title,
-                                imageName: inspectionDetailViewModel.imageName
-                            )
-                        }
-                        .listRowBackground(Color.clear)
-                        .listRowSeparator(.hidden)
-                    }
-                    .padding(.top, 100)
-                    .listStyle(.plain)
-                    
-                }
-//                .customNavigation/BarBackButtonHidden(false)
-            }
-            .tint(.black)
-            
+        VStack {
+            Text("ContentView")
         }
     }
 }
 
-#Preview {
-    ContentView()
+struct InformationView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView(title: "Title", mainText: "Main Text")
+    }
 }
