+++ CrimeFilesApp/CrimeFilesAppApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct CrimeFilesAppApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            MainView()
         }
     }
 }
