+++ CrimeFilesApp/FavoriteView.swift
@@ -1,0 +1,18 @@
+//
+//  FavoriteView.swift
+//  CrimeFilesApp
+//
+//  Created by Александр Полочанин on 8.11.23.
+//
+
+import SwiftUI
+
+struct FavoriteView: View {
+    var body: some View {
+        Text("FavoriteView")
+    }
+}
+
+#Preview {
+    FavoriteView()
+}
