+++ CrimeFilesApp/MainView.swift
@@ -1,0 +1,22 @@
+//
+//  MainView.swift
+//  CrimeFilesApp
+//
+//  Created by Александр Полочанин on 6.11.23.
+//
+
+import SwiftUI
+
+struct MainView: View {
+    
+    
+    var body: some View {
+        VStack {
+            Text("ContentView")
+        }
+    }
+}
+
+#Preview {
+    MainView()
+}
