+++ CrimeFilesApp/SectionView.swift
@@ -1,0 +1,18 @@
+//
+//  SectionView.swift
+//  CrimeFilesApp
+//
+//  Created by Александр Полочанин on 8.11.23.
+//
+
+import SwiftUI
+
+struct SectionView: View {
+    var body: some View {
+        Text("SectionView")
+    }
+}
+
+#Preview {
+    SectionView()
+}
