+++ CrimeFilesApp/SubSectionView.swift
@@ -1,0 +1,18 @@
+//
+//  SubSectionView.swift
+//  CrimeFilesApp
+//
+//  Created by Александр Полочанин on 8.11.23.
+//
+
+import SwiftUI
+
+struct SubSectionView: View {
+    var body: some View {
+        Text("SubSectionView")
+    }
+}
+
+#Preview {
+    SubSectionView()
+}
